Use vue router, add tabs in profile, move functionality to tabs, add some css vars
