+++ src/components/User/ProfileInfo.vue
@@ -1,0 +1,14 @@
+<template>
+  <p>ProfileInfo</p>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+  name: "ProfileInfo",
+});
+</script>
+
+<style scoped lang="scss">
+</style>
